handlers/start.py
from telegram.ext import CommandHandler, ContextTypes
from telegram import Update

async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    await update.message.reply_text("ברוך הבא! הבוט פעיל ומוכן.")

start_handler = CommandHandler("start", start)
